Rename DependencyInjection to DependencyRegistry
