--- src/dependency-injection/DependencyInjection.Tests/Test.DependencyRegistrationExtensions/DependencyRegistrationExtensionsTest.cs
@@ -1,8 +1,0 @@
-#nullable enable
-
-namespace PrimeFuncPack.Tests
-{
-    public sealed partial class DependencyRegistrationExtensionsTest
-    {
-    }
-}
