+++ src/dependency-registry/DependencyRegistry.Tests/DependencyInjection.Tests.csproj
@@ -12,7 +12,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="EarlyFuncPack.DependencyInjection" Version="*-*" />
+		<PackageReference Include="EarlyFuncPack.DependencyRegistry" Version="*-*" />
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.*" />
 		<PackageReference Include="Moq" Version="4.*" />
 		<PackageReference Include="PrimeFuncPack.UnitTest.Data" Version="2.0.*" />
