+++ src/dependency-registry/DependencyRegistry.Tests/Mock/MockServiceCollection.cs
