+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistrar/DependencyRegistrarTest.cs
