+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistrar/RegistrarTest.Factory.cs
