+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistrar/RegistrarTest.RegScoped.cs
