+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistrar/RegistrarTest.RegSingleton.cs
