+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistrar/RegistrarTest.RegTransient.cs
