+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistryExtensions/DependencyRegistryExtensionsTest.cs
@@ -1,0 +1,8 @@
+#nullable enable
+
+namespace PrimeFuncPack.Tests
+{
+    public sealed partial class DependencyRegistryExtensionsTest
+    {
+    }
+}
