+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistryExtensions/ExtensionsTest.RegScoped.cs
@@ -9,38 +9,38 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Tests
 {
-    partial class DependencyRegistrationExtensionsTest
+    partial class DependencyRegistryExtensionsTest
     {
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public void RegisterSingleton_ExpectSourceServices(
+        public void RegisterScoped_ExpectSourceServices(
             bool isNotNull)
         {            
             var mockServices = MockServiceCollection.CreateMock();
             var sourceServices = mockServices.Object;
             
-            RefType regService = isNotNull ? MinusFifteenIdRefType : null!;
+            RefType regService = isNotNull ? ZeroIdRefType : null!;
             var dependency = Dependency.Create(_ => regService);
 
             var registrar = dependency.ToRegistrar(sourceServices);
 
-            var actualServices = registrar.RegisterSingleton();
+            var actualServices = registrar.RegisterScoped();
             Assert.Same(sourceServices, actualServices);
         }
 
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public void RegisterSingleton_ExpectCallAddSingletonOnce(
+        public void RegisterScoped_ExpectCallAddScopedOnce(
             bool isNotNull)
         {
-            object regService = isNotNull ? MinusFifteenIdNullNameRecord: null!;
+            RecordType regService = isNotNull ? MinusFifteenIdSomeStringNameRecord: null!;
             var mockServices = MockServiceCollection.CreateMock(
                 sd =>
                 {
-                    Assert.Equal(typeof(object), sd.ServiceType);
-                    Assert.Equal(ServiceLifetime.Singleton, sd.Lifetime);
+                    Assert.Equal(typeof(RecordType), sd.ServiceType);
+                    Assert.Equal(ServiceLifetime.Scoped, sd.Lifetime);
                     Assert.NotNull(sd.ImplementationFactory);
 
                     var actualService = sd.ImplementationFactory!.Invoke(Mock.Of<IServiceProvider>());
@@ -51,7 +51,7 @@ namespace PrimeFuncPack.Tests
             var dependency = Dependency.Create(_ => regService);
 
             var registrar = dependency.ToRegistrar(sourceServices);
-            _ = registrar.RegisterSingleton();
+            _ = registrar.RegisterScoped();
             
             mockServices.Verify(s => s.Add(It.IsAny<ServiceDescriptor>()), Times.Once);
         }
