+++ src/dependency-registry/DependencyRegistry.Tests/Test.DependencyRegistryExtensions/ExtensionsTest.RegSingleton.cs
@@ -9,49 +9,49 @@ using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Tests
 {
-    partial class DependencyRegistrationExtensionsTest
+    partial class DependencyRegistryExtensionsTest
     {
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public void RegisterTransient_ExpectSourceServices(
+        public void RegisterSingleton_ExpectSourceServices(
             bool isNotNull)
         {            
             var mockServices = MockServiceCollection.CreateMock();
             var sourceServices = mockServices.Object;
             
-            RecordType regService = isNotNull ? ZeroIdNullNameRecord : null!;
+            RefType regService = isNotNull ? MinusFifteenIdRefType : null!;
             var dependency = Dependency.Create(_ => regService);
 
             var registrar = dependency.ToRegistrar(sourceServices);
 
-            var actualServices = registrar.RegisterTransient();
+            var actualServices = registrar.RegisterSingleton();
             Assert.Same(sourceServices, actualServices);
         }
 
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
-        public void RegisterTransient_ExpectCallAddTransientOnce(
+        public void RegisterSingleton_ExpectCallAddSingletonOnce(
             bool isNotNull)
         {
-            string regService = isNotNull ? LowerSomeString: null!;
+            object regService = isNotNull ? MinusFifteenIdNullNameRecord: null!;
             var mockServices = MockServiceCollection.CreateMock(
                 sd =>
                 {
-                    Assert.Equal(typeof(string), sd.ServiceType);
-                    Assert.Equal(ServiceLifetime.Transient, sd.Lifetime);
+                    Assert.Equal(typeof(object), sd.ServiceType);
+                    Assert.Equal(ServiceLifetime.Singleton, sd.Lifetime);
                     Assert.NotNull(sd.ImplementationFactory);
 
                     var actualService = sd.ImplementationFactory!.Invoke(Mock.Of<IServiceProvider>());
                     Assert.Equal(regService, actualService);
                 });
 
-            var sourceServices = mockServices.Object;            
+            var sourceServices = mockServices.Object;
             var dependency = Dependency.Create(_ => regService);
 
             var registrar = dependency.ToRegistrar(sourceServices);
-            _ = registrar.RegisterTransient();
+            _ = registrar.RegisterSingleton();
             
             mockServices.Verify(s => s.Add(It.IsAny<ServiceDescriptor>()), Times.Once);
         }
