+++ src/dependency-registry/DependencyRegistry/DependencyRegistrar.T/DependencyRegistrar.T.cs
