+++ src/dependency-registry/DependencyRegistry/DependencyRegistrar.T/RegisterScoped.cs
