+++ src/dependency-registry/DependencyRegistry/DependencyRegistrar.T/RegisterSingleton.cs
