+++ src/dependency-registry/DependencyRegistry/DependencyRegistrar/DependencyRegistrar.cs
