+++ src/dependency-registry/DependencyRegistry/DependencyRegistry.csproj
@@ -10,7 +10,7 @@
 		<Description>PrimeFuncPack: A Functional Programming Pack for .NET</Description>
 		<Copyright>Copyright © 2020 Andrei Sergeev, Pavel Moskovoy</Copyright>
 		<RootNamespace>PrimeFuncPack</RootNamespace>
-		<AssemblyName>EarlyFuncPack.DependencyInjection</AssemblyName>
+		<AssemblyName>EarlyFuncPack.DependencyRegistry</AssemblyName>
 		<Version>0.0.1</Version>
 	</PropertyGroup>
